+++ tests/test_tables.py
@@ -1,4 +1,5 @@
 import pytest
+import re
 
 from tinydb import where
 
@@ -44,6 +45,7 @@ def test_caching(db):
     table2 = db.table('table1')
 
     assert table1 is table2
+
 
 def test_zero_cache_size(db):
     table = db.table('table3', cache_size=0)
@@ -104,3 +106,14 @@ def test_table_name(db):
 
     with pytest.raises(AttributeError):
         table.name = 'foo'
+
+
+def test_table_repr(db):
+    name = 'table4'
+    table = db.table(name)
+    print(repr(table))
+
+    assert re.match(
+        r"<Table name=\'table4\', total=0, "
+        "storage=<tinydb\.database\.StorageProxy object at [a-zA-Z0-9]+>>",
+        repr(table))